+++ packages/account-api-sdk/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sophon-labs/account-api-sdk
 
+## 3.0.0-next.37
+
+### Patch Changes
+
+- adding signer to biconomy calls
+- Updated dependencies
+  - @sophon-labs/account-core@3.0.0-next.37
+
 ## 3.0.0-next.36
 
 ### Patch Changes